Create h2o first Attempt
This was the first attempt at running Gradient Boosting Machines on the data. I also tried to run an Auto ML - but I did not understand the output, so I said it didn't work. I realized after a couple of days that it actually had run properly. Really was just trying to understand the system with this first attempt. This script got me a final score of 51.