Create h2o Model 3 (GBM - Random Forest)
This was the 2nd Model that I tried to run on h2o. Georgia had talked about getting good results with random forests so I decided to attempt that on a variety High_Performer and Retained. I also looked into Deep Learning from speaking to my friend that I was working on the project with. This model gave me many low Final scores (into the negatives).