Create h2o Model 4 - Protected Var Importance
What I did with this model was a little bit different. I had run a model in model three that gave me the importance of variables with the outcome variable of Protected_Group. I decided that I would only include the variables that had standardized importance of above .1. I ran an AutoML with the outcome variable Retained and got a 49 and then ran an AutoML with outcome variable High_Performer that got me a 59 Final Score. I tried to mix predictions from a variety of models that I had run and the best was the AutoML with High_Performer. I used this AutoML model to predict on the final data but got a poor/low score of 15.