Create R Code for h2o
This was the r code that I had to run to be able to access h2o. I also brought some of the code from h2o into r to be able to keep track of which models I was testing. Included are some of the results for the machine learning competition.